Initial commit
It is ready to be run in any browser "Developper Tools" console and is compatible with Infinite Craft Helper recipes and Pinned. Just needs a discord WebHook if you like the recipes to be stored in a discord channel. By default It will try to save a JSON file in every loop of a newly discovered recipe, optionally you could disable it. It only stores, logs only elements with valid English or Greek words, chacked against neurolingo.gr project api. USE RESPONSIBLY. Credits where it is due: to the reddit user @Eva-Rosalene that was my initial inspiration and of course to @Mikarific ,the developer of Infinite Craft Helper that is doing an awesome job making the game so much fun to play, and finally also to nata, the developer of Mongolian Beef Script!